Handle directories with odd names
Previously, adding a directory with square brackets
resulted in a strange, silent error [1]. This happened
because directory names were being passed to a file-
matching glob designed to recursively stream files from
the directory. When filenames contained meaningful glob
symbols (like [ or *), the glob matching would fail.

This commit resolves this problem by escaping glob
characters on filenames, before the filename is inserted
into the main glob in src/get-file-streams.js L48.
We add a dependency to glob-escape in the process,
and a new test in test/ipfs-api/util.spec.js.