Change default config from JSON file to JS module
Because the default config was a JSON file, it was treated as a singleton object. In some parts of the code, data got added to it, leading to race conditions when creating multiple nodes simultaneously. To make that harder to do accidentally, the default config is now a *function* that returns a unique configuration object on each call.

Fixes #1316.

License: MIT
Signed-off-by: Rob Brackett <rob@robbrackett.com>