fix: configure webpack to not use esmodules in dependencies
Big.js just released an update that added esmodule support by adding a "module" field to their package.json. By default, webpack will use this field when requiring a module. Since esmodules aren't fully supported in the browser and we don't do any transpiling this PR configures webpack to not consider the esmodule field.

[Webpack `mainFields` docs](https://webpack.js.org/configuration/resolve/#resolve-mainfields)

[Why `require('big.js')` is broken for us](https://github.com/webpack/webpack/issues/4742)

Refs https://github.com/ipfs/js-ipfs/issues/1363

Note that this solves the issue for `big.js`, but also for any other modules in the future that decide to define a `module` field in their `package.json` or any that already do and we just haven't realised it yet!

License: MIT
Signed-off-by: Alan Shaw <alan@tableflip.io>