fix: emit boot error only once
`init` and `start` both emit any error they encounter. `boot` will then receive that error and also emit it! This PR adds an `emitted` property in `boot` to errors that came from `init` or `start` so that later in the code it knows whether to emit it or not!

License: MIT
Signed-off-by: Alan Shaw <alan@tableflip.io>