fix: report correct size for raw dag nodes
If you specify `cidVersion: 0`, for go compatibility you will have raw leaf nodes.
If your data fits in one dag node, the root node will be a raw node for the same
reason, and `ipfs.object.get` will return a buffer.

This PR inspects the returned node to see if it's a buffer, if it is, use the
`length` property, otherwise use the DAGNode `size` property.

Fixes #1585