test: remove dupe api tests
The http-api test suite was doing redundant work. The removed
tests were spinning up a daemon to run the http-api tests against.
This duplicates the behavior the cli tests are already doing. Also,
the inject http-api tests, test the api directly, without spinning up
a daemon. This is a better approach and makes these deleted tests
redundant and useless. This also shaves off a significant time from
the test suite.

Resolves https://github.com/ipfs/js-ipfs/issues/1666