feat: upload example works with big files
Old example created a Buffer from the file handle which quickly hit
allocation limits with big files.

This change replaces Buffer with a stream, which enables example to work
with big files (tested with 3.6GB video, took ~30seconds with go-ipfs on
localhost).

While I was at it, I added a demo of a pattern for persisting filenames
by wrapping them in unixfs directory.

License: MIT
Signed-off-by: Marcin Rataj <lidel@lidel.org>