fix: skip test that go-ipfs cannot pass
For some reason go-ipfs does not support a `version` parameter in the `/api/v0/block/put` endpoint but js-ipfs does. This means that go-ipfs cannot pass this skipped CID version agnostic test. When IPFS switches to v1 CIDs by default we'll have to switch this skipped test to the one that tests the other way around i.e. CIDv0 to CIDv1.

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>