fix: path to cid-tool commands (#1866)
In the edge case where IPFS is a dependency of a project and the user attempts to use the IPFS CLI, the cid-tool will not be installed at `proj/node_modules/ipfs/node_modules/cid-tool` but will instead be flattened by npm and installed at `proj/node_modules/cid-tool`. i.e. next to `ipfs` in the dependency tree instead of a child of.

This PR fixes this situation by using `require.resolve` to get the path to the `cid-tool` instead of assuming where it is installed.

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>