fix: cat deeply nested file
The exporter was exporting the file, but the filter was filtering it out because the exported path includes only `file.txt` but the filter was expecting `path/file.txt`.

So when you cat `/ipfs/QmHash/file.txt` it worked, but when you cat `/ipfs/QmHash/path/file.txt` it did not.

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>