fix: update peer info before conditionally dialing
When we discover a peer, we may not have the latest information from our PeerBook, such as the connected multiaddress. Doing peerBook.put handles merging this information and the new PeerInfo instance is returned. Updating the peerInfo instance, before performing a check, should make sure we have the correct information.

Co-Authored-By: alanshaw <alan.shaw@protocol.ai>