refactor: swap joi-browser with superstruct (#1961)
Relay config is removed because libp2p already validates it. A bug was detected and fixed because the validation is stricter. This should improve the overall bundle size and safety handling the config.

BREAKING CHANGE: Constructor config validation is now a bit more strict - it does not allow `null` values or unknown properties.