fix: stop IPNS asap
This PR refactors the stop logic to perform stop actions in parallel for various components (where it is safe to do so). This means that libp2p gets stopped at the same time as the republisher. When libp2p stops it closes all connections meaning that the republsher is able to complete a running republish sooner.

Before, we had to wait for republish to complete before stop happens (since there's no way to cancel a DHT put right now).

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>