fix: Do not load all of a DAG into memory when pinning
Given a `CID`, the `dag._recursiveGet` method returns a list of all
descendents of the node with the passed `CID`.  This can cause
enormous memory useage when importing large datasets.

Where this method is invoked the results are either a) disgarded or
b) used to calculate the `CID`s of the nodes which is then bad for memory
*and* CPU usage.

This PR removes the buffering and `CID` recalculating for a nice speedup
when adding large datasets.

fixes #2310