fix: make invalid url actually invalid
I'm using one of those ISPs that 'helpfully' directs you to a search
page if the URL you are accessing does not have a DNS match.

The URL used in the test was well formed but not a real domain though
can be resolved, so this PR changes it to be something even a
web-breaking DNS server would refuse to resolve.