fix: pull in preconfigured chai from interface tests
This is to work around https://github.com/chaijs/chai/issues/1298

Fundamentally we cannot pull in chai and add plugins to it without
being careful of the order of those plugins as it's a singleton.

All we're really interested in is the `expect` function to just
get the preconfigured one from the interface tests.