perf: debounce pin flush
When importing a directory of files, after every block is added we
pin it and flush pins to the repo.  This is really slow and if we
are adding lots of files, the pinset is out of date almost
immediately so this PR debounces flushing pins until after we've
stopped adding things.

A call to `pinManager.flushPins` will return a promise, and that
promise will resolve once the pins have been flushed, but it will
first wait 100ms and see if any subsequent flushes have been
requested in that time - if so that promise will end early and
return the promise created by the later flush request.