chore: split test jobs into stages
The current test stage uses _all_ the workers we have available in Travis and starves other projects from running their tests. This change separates interop, externals and examples into stages that only execute if the previous was successful. This means that we use fewer workers concurrently and fewer workers overall since a failure in the main tests will not process and run interop/externals and examples tests. This is at the expense of slower overall CI.

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>