chore: allow externals to fail
We need to publish an rc for people to use to refactor their code
to use the new API.

The `aegir` release commands require the `last-successful` build
to have run in order to update the last known good build branch.

That build will not run until the external builds pass but we can't
make them pass until they can be refactored to use the new API for
which we need an rc.

The change to the `.travis.yml` in this PR allows the overall build
to pass even though the external module builds are faiilng.