fix: print peer id in swarm addresses on startup
Since https://github.com/libp2p/js-libp2p/pull/558 js-libp2p now
requires peer ids to be part of dialled multiaddrs.

Previously a user was able to copy/paste swarm addresses printed as
part of the daemon starting up but now they can't because the peer
id is not included in the console output.

This PR prints the node's peer id as part of the swarm address to
let people continue doing this and fixes the custom-libp2p example.