block.rm: remove second greaterThanZero check
This PR removes the `expect(localRefsAfterRemove).to.have.property('length').that.is.greaterThan(0)` from line 42. The rationale here is that the `rs-ipfs` nodes start with a completely empty blockstore. Once only block added is removed, then it drops back to one again.

This was imo the least obtrusive move I could make with the tests, the alternative could be to seed a cid2 within the test to ensure it was always `greaterThanZero`