fix: allow either cidv{0,1} in refs-local test
rationale: simplest way to support using cidv0 and cidv1 interchangeably
in an ipfs implementation is to just convert all cids to cidv1
transparently. this has the miniscule setback of `refs/local` not being able
to list cids in their original version. this fix allows the
implementation to use either version internally.