chore: disable codecov
It misreports coverage numbers because it doesn't take into account
that only the tests of changed modules and their dependents run in
each build, it only reports coverage of node tests, it marks every
PR as failed build checks yet we merge them anyway.

We can re-enable it when we solve the problems above.