refactor(tests): replace ipfs.add in cat tests (#3078)
Continuing the trend of replacing ipfs.add use in the tests, this time with cat tests. This follows my previous PR's and PR's by @aphelionz.

This is currently a draft as I did not yet add path handling in my parallel rust-ipfs work so those test cases are failing quite bad at the moment. So I don't know if it's because of the unimplemented path support or because I messed something up in the test code.