docs: ready to test in less time
This is an improvement to a previous PR (#3084 ). It is ment to avoid building all the project just to run an example. Following the old instructions, after the cloning is finished, it took me 20 minutes before I can run the example. With this modification, I can run the example just 2 minutes after the cloning is finished.