fix: still load dag-pb, dag-cbor and raw when specifying custom formats
If we specify a `formats` array as part of the ipld options in in-proc
nodes, it replaces the default list of dag-pb, dag-cbor and raw.

This change allows the `loadFormat` function to still resolve those
formats even if the user has passed a `format` array that does not
contain them.

Fixes #3129