fix: normalise add input to blob in the browser
To support streaming of native types with no buffering, normalise add
input to async iterators in node and Blobs in the browser.

That is, if the user passes a blob in the browser leave it alone as
enumerating blob contents cause the files to be read.

Browser FormData objects do not allow you to specify headers for each
multipart part which means we can't pass unixfs metadata via the headers
so we turn the metadata into a querystring and append it to the field
name for each multipart part as a workaround.

BREAKING CHANGES:

- Removes the `mode`, `mtime` and `mtime-nsec` headers from multipart requests
- Passes `mode`, `mtime` and `mtime-nsec` as querystring parameters appended to the field name of multipart requests