chore: consistent browser examples (#3204)
This PR aims to turn browser examples naming consistent. Only the directories were renamed to be prefixed by `browser-*` like the other examples. I was always trying to find the exchange files example in the browser and I was always wrongly directioned by the `browser-*` naming, where it was not.

Co-authored-by: achingbrain <alex@achingbrain.net>