chore: run docker rc release separately
When we publish an rc to npm, sometimes the release isn't available
by the time we try to publish an rc to docker hub.

Change the docker publish to be a separate job to give it a bit of
time before running so npm has a chance to propagate the rc but also
so we can re-run if it's too fast.