feat: remove esoteric ipld formats
We bundle ipld formats for ethereum, zcash, bitcoin and git though they
speak beyond the files part of the Interplanetary File System.

Since #3347 we can now configure extra IPLD formats in the http client,
in-process and daemon nodes, we no longer have to bundle these formats
for them to be supported, instead the user can choose to configure their
node with the formats they require.

This makes the behaviour of core the same in node as it is in the browser
and also means we don't waste time installing deps our users may not use.

If they do use them, they can configure the node as they see fit.

BREAKING CHANGE: only dag-pb, dag-cbor and raw formats are supported out
  of the box, any others will need to be configured during node startup.