chore: automate updating example dependencies
Due to not wanting to publish versions of examples on GH or npm we
don't include the examples in our lerna workspace during publishing.

Consequently the `package.json` of examples doesn't get updated with
new `ipfs` etc modules during the publish process, which means they
depend on old versions and can break, so we need to update the versions
manually which is tiresome and error prone.

I've added a script to do it at the end of the publish process.