feat: add new BlockService
This commits adds the BlockService directly to IPFS. It takes js-multiformats
CIDs as input and also returns new style blocks and no longer IpldBlocks.

It currently converts those CIDs and blocks into the legacy format, in order
to be compatible with the current ipfs-repo and ipfs-bitswap implementations.
In the future this kind of conversion will disappear, once the full stack is
using js-multiformats.