chore: fix docker images
The python package has gone, we need to use python3 instead.

There's a futher problem with npm not setting the $PATH properly so
it fails as it can't find pbjs, but this is only necessary with gh
dep versions which should go away with the next libp2p-kad-dht/libp2p-gossipsub
releases.