fix: reject requests when cors origin list is empty
If CORS origin list is empty, Hapi throws an error as it considers
that to be invalid configuration.

We want to reject requests that send and origin or a referrer when
no allowed origins have been configured, so when these headers are
sent, reject the request if no allowed origins are present in the
config.