chore: allow bundlesize gh action to fail
Sometimes it's because there's a duplicate dep version that needs to
be fixed.

Usually it's because some transitive dep increases it's bundle size
by a kilobyte or two and has nothing to do with the current PR, so all
we do is increase the bundlesizeMax value to make it pass.

Let's just make the check informational instead.