fix: move wrtc to optional deps (#3705)
npm7 has started installing peer deps by default, which means if you
are running a future tech node.js version you may end up compiling
native addons that may fail.  These addons should be opt-in so move
them to optional deps which means if they fail to install, the whole
installation does not fail.

Also adds `@mapbox/node-pre-gyp` as a dependency so we can attempt to 
compile modules that have native addons where no prebuilt binary is
available instead of requiring users to install it globally which is
required due to a misconfiguration in the `webrtc` module.