fix: await pubsub unsubscribe to fix flaky test
We manage pubsub subscriptions locally in the http client, when we unsubscribe
from a given topic, we abort the long-lived connection we're holding open but
do not wait for it to actually close.

In our tests we unsubscribe then check the subscription list - if we do this
before the subscription is torn down it can report that we are still subscribed
to the topic we just unsubscribed from.

The fix here is to wait for the connection to be torn down before returning
from the unsubscribe call, by which point we should no longer have the topic
in our subs list.