feat: make ipfs.get output tarballs (#3785)
`ipfs get` on the cli writes files/folders to disk, also tarballs
and gzipped tarballs. Also gzipped individual files.

In the API it's very similar to `ipfs.ls`.  Make it more like the
cli so there's a clear reason to use one or the other.

Also adds types to `interface-ipfs-core` to better catch broken
types in `ipfs-core-types`.

BREAKING CHANGE: the output type of `ipfs.get` has changed and the `recursive` option has been removed from `ipfs.ls` since it was not supported everywhere