fix: override node-fetch version (#3835)
We need the fix from https://github.com/node-fetch/node-fetch/pull/1172
in node-fetch otherwise we can't detect when the other end of a streaming
response closes the connection.

At the moment we're using a fork of node-fetch with this patch applied
but the npm7 depdending hoisting algorithm hoists the wrong version so
here we add a dep to the root package.json to ensure we get the right
one.

This should only affect development and not published versions.  We can
revert this and switch to undici when node 16 becomes lts at the end
of October.